pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "A simple project to prepare data for mlops project"
authors = ["mdop297"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
hydra-core = "~=1.3.2"
pydantic = "~=2.9.2"
pandas = "~=2.2.3"
nltk = "~=3.9"
symspellpy = "~=6.7"
fsspec = {version = "~=2024.9", extras = ["gcs"]}
gcsfs = "~=2024.9"
google-cloud-secret-manager = "~=2.16"
fastparquet = "~=2024.5"
dvc = {version = "~=3.55.2", extras = ["gdrive", "gs"]}
dask = {version = "~=2024.9.1", extras = ["complete"]}
distributed = "~=2024.9.1"
dask-cloudprovider = {version = "2024.9.1", extras=["gcp"]}
dask-ml = "~=2024.4"
bokeh = "~=3.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "~=8.3.3"
black = "~=24.8.0"
isort = "~=5.13.2"
flake8 = "~=7.1.1"
mypy = "~=1.11.2"
jupyterlab = "~=4.2.5"
types-PyYAML = "~=6.0.12.20240917" 
types-cachetools = "~=5.5.0.20240820"
types-decorator = "~=5.1.8.20240310"
types-protobuf = "~=5.28.0.20240924"
types-pycurl = "~=7.45.3.20240421"
types-requests = "~=2.32.0.20240914"
types-six = "~=1.16.21.20240513"

[tool.poetry.group.docs.dependencies]
sphinx = "~=8.0.2"

[tool.poetry.group.docs]
optional = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["cybulde"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
